requirements.txt
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==2.0.0
configparser==7.2.0
cssselect2==0.9.0
defusedxml==0.7.1
packaging==26.0
pillow==12.1.1
pycparser==3.0
pywayland==0.4.18
pyxdg==0.28
tinycss2==1.5.1
webencodings==0.5.1
xdg==6.0.0
